Split out the IfcHandler
