+++ src/Makefile
@@ -86,8 +86,9 @@ SPROUT_COMMON_SOURCES := logger.cpp \
                          base64.cpp \
                          as_communication_tracker.cpp \
                          astaire_resolver.cpp \
-                         reg_data_xml_utils.cpp \
-                         wildcard_utils.cpp
+                         xml_utils.cpp \
+                         wildcard_utils.cpp \
+                         ifc.cpp
 
 sprout_SOURCES := ${SPROUT_COMMON_SOURCES} \
                   snmp_counter_table.cpp \
